Récupération des Micro Mesures et mise à jour de la récupération des Mesures
La récupération des Mesures et des Micro Mesures est possible au utilisateur les objets de mesures ce présente sous la forme d'un nom, d'un attribut qui est celui que le code reconnais ainsi que d'une description vide pour l'instant

+++ src/main/java/com/lds_api/service/LDSService.java
@@ -106,7 +106,7 @@ public class LDSService {
 			double score = 0.0;
 			
 			if(edges1.size() == 1 && edges2.size() == 1) {
-				if(params.getOptions().getMeasureType().equals("int")) {
+				if(params.getOptions().getMeasureType().equals("numeric")) {
 					String[] tab1 = edges1.get(0).split("\"");
 					String[] tab2 = edges2.get(0).split("\"");
 					score = IntMeasure(Double.parseDouble(tab1[1]),Double.parseDouble(tab2[1]),params.getOptions().getMeasureType());
